+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/registry/ModFeatures.kt
@@ -1,9 +1,9 @@
 package dev.aaronhowser.mods.excessive_utilities.registry
 
 import dev.aaronhowser.mods.excessive_utilities.ExcessiveUtilities
-import dev.aaronhowser.mods.excessive_utilities.feature.DeepDarkPillarFeature
-import dev.aaronhowser.mods.excessive_utilities.feature.DeepDarkSculkVeinFeature
-import dev.aaronhowser.mods.excessive_utilities.feature.DeepDarkStalactiteFeature
+import dev.aaronhowser.mods.excessive_utilities.feature.DepthsPillarFeature
+import dev.aaronhowser.mods.excessive_utilities.feature.DepthsSculkVeinFeature
+import dev.aaronhowser.mods.excessive_utilities.feature.DepthsStalactiteFeature
 import net.minecraft.core.registries.BuiltInRegistries
 import net.minecraft.world.level.levelgen.feature.Feature
 import net.neoforged.neoforge.registries.DeferredHolder
@@ -14,13 +14,13 @@ object ModFeatures {
 	val FEATURE_REGISTRY: DeferredRegister<Feature<*>> =
 		DeferredRegister.create(BuiltInRegistries.FEATURE, ExcessiveUtilities.MOD_ID)
 
-	val DEEP_DARK_PILLAR: DeferredHolder<Feature<*>, DeepDarkPillarFeature> =
-		FEATURE_REGISTRY.register("deep_dark_pillar", ::DeepDarkPillarFeature)
+	val DEPTHS_PILLAR: DeferredHolder<Feature<*>, DepthsPillarFeature> =
+		FEATURE_REGISTRY.register("depths_pillar", ::DepthsPillarFeature)
 
-	val DEEP_DARK_STALACTITE: DeferredHolder<Feature<*>, DeepDarkStalactiteFeature> =
-		FEATURE_REGISTRY.register("deep_dark_stalactite", ::DeepDarkStalactiteFeature)
+	val DEPTHS_STALACTITE: DeferredHolder<Feature<*>, DepthsStalactiteFeature> =
+		FEATURE_REGISTRY.register("depths_stalactite", ::DepthsStalactiteFeature)
 
-	val DEEP_DARK_SCULK_VEIN: DeferredHolder<Feature<*>, DeepDarkSculkVeinFeature> =
-		FEATURE_REGISTRY.register("deep_dark_sculk_vein", ::DeepDarkSculkVeinFeature)
+	val DEPTHS_SCULK_VEIN: DeferredHolder<Feature<*>, DepthsSculkVeinFeature> =
+		FEATURE_REGISTRY.register("depths_sculk_vein", ::DepthsSculkVeinFeature)
 
 }